NEXUS-23434: Refactor Orient/StorageTX-isms from Helm
