+++ nexus-repository-helm/src/main/java/org/sonatype/repository/helm/internal/orient/HelmComponentMaintenanceFacet.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.helm.internal;
+package org.sonatype.repository.helm.internal.orient;
 
 import java.util.Collections;
 import java.util.Set;
