+++ nexus-repository-helm/src/main/java/org/sonatype/repository/helm/internal/orient/createindex/CreateIndexService.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.helm.internal.createindex;
+package org.sonatype.repository.helm.internal.orient.createindex;
 
 import org.sonatype.nexus.repository.Repository;
 import org.sonatype.nexus.repository.storage.TempBlob;
