+++ nexus-repository-helm/src/main/java/org/sonatype/repository/helm/internal/orient/hosted/HelmHostedFacet.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.helm.internal.hosted;
+package org.sonatype.repository.helm.internal.orient.hosted;
 
 import java.io.IOException;
