+++ nexus-repository-helm/src/main/java/org/sonatype/repository/helm/internal/orient/metadata/IndexYamlAbsoluteUrlRewriter.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.helm.internal.metadata;
+package org.sonatype.repository.helm.internal.orient.metadata;
 
 import java.io.IOException;
 import java.io.InputStream;
