+++ nexus-repository-helm/src/main/java/org/sonatype/repository/helm/internal/util/TgzParser.java
@@ -13,7 +13,6 @@
 package org.sonatype.repository.helm.internal.util;
 
 import java.io.ByteArrayInputStream;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
