+++ nexus-repository-helm/src/test/resources/org/sonatype/repository/helm/internal/orient/metadata/index.yaml
